+++ web/documentserver-example/java-spring/src/main/resources/static/scripts/jscript.js
@@ -142,7 +142,7 @@ if (typeof jQuery !== "undefined") {
                     var responseText = data.responseText;
                     var response = jq.parseJSON(responseText);
                     if (response.error) {
-                        if (response.error.includes("Incorrect password")) {
+                        if (response.error == "PASSWORD") {
                             jq(".current").removeClass("current");
                             jq("#step2").addClass("error");
                             jq("#blockPassword").show();
@@ -152,7 +152,7 @@ if (typeof jQuery !== "undefined") {
                             }
                             return;
                         } else {
-                            if (response.error.includes("Error conversion output format")){
+                            if (response.error == "OOXML_OUTPUT_TYPE"){
                                 jq("#select-file-type").removeClass("invisible");
                                 jq("#step2").removeClass("current");
                                 jq("#hiddenFileName").attr("placeholder",filePass);
@@ -160,17 +160,16 @@ if (typeof jQuery !== "undefined") {
                             }
                             jq(".current").removeClass("current");
                             jq(".step:not(.done)").addClass("error");
-                            jq("#mainProgress .error-message").show().find("span").text(response.error);
+                            jq("#mainProgress .error-message").show().find("span").text("Error automatically determine the output file format");
                             jq('#hiddenFileName').val("");
                             return;
                         }
                     }
 
-                    jq("#hiddenFileName").val(response.filename);
-
-                    if (response.step && response.step < 100) {
+                    if (response.hasOwnProperty("percent") && response.percent < 100) {
                         checkConvert(filePass, fileType);
                     } else {
+                        jq("#hiddenFileName").val(response.filename);
                         jq("#step2").addClass("done").removeClass("current");
                         loadScripts();
                     }
@@ -430,4 +429,27 @@ if (typeof jQuery !== "undefined") {
         jq("div.tooltip").remove();
     });
 
+}
+
+function toggleSidePanel(event) {
+    event.preventDefault();
+    let sidePanel = document.querySelector(".left-panel");
+    let body = document.querySelector("body");
+    if (sidePanel.classList.contains("active")) {
+        sidePanel.classList.remove("active");
+        body.classList.remove("menu-open");
+    } else {
+        sidePanel.classList.add("active")
+        body.classList.add("menu-open");
+    }
+}
+
+function toggleUserDescr(event) {
+    let list = event.currentTarget.querySelector("ul");
+    let cursor = window.getComputedStyle(event.currentTarget).getPropertyValue("cursor");
+
+    if (cursor === "pointer") {
+        if (list.classList.contains("active")) list.classList.remove("active");
+        else list.classList.add("active");
+    }
 }