+++ web/documentserver-example/nodejs/helpers/wopi/wopiRouting.js
@@ -49,9 +49,11 @@ exports.registerRoutes = function registerRoutes(app) {
     const actions = await utils.getDiscoveryInfo(req.DocManager);
     const wopiEnable = actions.length !== 0;
     const docsExtEdit = []; // Supported extensions for WOPI
+    const editNewExts = [];
 
     actions.forEach((el) => {
       if (el.name === 'edit') docsExtEdit.push(`${el.ext}`);
+      if (el.name === 'editnew') editNewExts.push({ ext: el.ext, text: utils.getEditNewText(el.ext) });
     });
 
     // Checking supported extensions
@@ -71,7 +73,16 @@ exports.registerRoutes = function registerRoutes(app) {
         file.actions = await utils.getActions(req.DocManager, ext); // get actions of the specified extension
         // eslint-disable-next-line no-await-in-loop
         file.defaultAction = await utils.getDefaultAction(req.DocManager, ext);// get the default action for extension
-        if (mobile) file.actions.forEach((act) => { if (act.name === 'mobileEdit') file.defaultAction = act; });
+        if (mobile) {
+          file.actions.forEach((act) => {
+            if ((file.defaultAction.name === 'edit' || file.defaultAction.name === 'formsubmit')
+              && act.name === 'mobileEdit') {
+              file.defaultAction = act;
+            } else if (file.defaultAction.name === 'view' && act.name === 'mobileView') {
+              file.defaultAction = act;
+            }
+          });
+        }
       }
 
       // render wopiIndex template with the parameters specified
@@ -85,6 +96,8 @@ exports.registerRoutes = function registerRoutes(app) {
         editedExts,
         fillExts,
         languages: configServer.get('languages'),
+        enableForgotten: configServer.get('enableForgotten'),
+        editNewExts,
       });
     } catch (ex) {
       console.log(ex); // display error message in the console