+++ web/documentserver-example/ruby/app/configuration/configuration.rb
@@ -31,7 +31,7 @@ class ConfigurationManager
 
   sig { void }
   def initialize
-    @version = '1.8.0'
+    @version = '1.10.0'
   end
 
   sig { returns(T.nilable(URI::Generic)) }
@@ -114,6 +114,14 @@ class ConfigurationManager
     false
   end
 
+  sig { returns(T::Boolean) }
+  def enable_forgotten
+    env = ENV.fetch('ENABLE_FORGOTTEN', nil)
+    return ActiveModel::Type::Boolean.new.cast(env) if env
+
+    true
+  end
+
   sig { returns(Pathname) }
   def storage_path
     storage_path = ENV['STORAGE_PATH'] || 'storage'
@@ -174,7 +182,8 @@ class ConfigurationManager
       'pt-PT': 'Portuguese (Portugal)',
       ro: 'Romanian',
       ru: 'Russian',
-      'sr-Latn-RS': 'Serbian',
+      'sr-Cyrl-RS': 'Serbian (Cyrillic)',
+      'sr-Latn-RS': 'Serbian (Latin)',
       si: 'Sinhala (Sri Lanka)',
       sk: 'Slovak',
       sl: 'Slovenian',